Add icons. Add day/night/lightning. Country vars
Added an Image file and animated and static weather
images.

Added elements for them in html

Changed the query parameters from API to include
day and night and the chance of lightning.

Gave variables to the two countries for clarity when reading
Added a catch for if the data is invalid.

Commented logic for icons nad started to apply code.